requirements.txt
Flask==2.0.1
Flask-MySQLdb==1.0.1
python-dotenv==0.19.0
werkzeug==2.0.1
reportlab==3.6.12